Add more details to deploy/README.md
In the process of creating a local installation, I have found additional
requirements and also modified the location of some configurations. Here
I describe these findings and changes.

Also I take this opportunity to describe how to use the local
installation. I hope this will be useful for the development of the
deployment scripts here contained.