Deploy a simple helloworld stack to Swarm
The deployed service is simple, but this is a important change as it
sets the base for the deployment of more complex stacks.

The trick part here is that the `docker_stack` ansible module performs
its tasks based on the host environment and filesystem. This has some
consequences:

* the compose file must be copied to the host filesystem
* python dependencies must be installed on the host

More on this module at:
https://docs.ansible.com/ansible/latest/modules/docker_stack_module.html.