Use ansible module for adding docker apt key
This is the way to perform such task on ansible. The key ID is the one
specified in: https://docs.docker.com/install/linux/docker-ce/debian/.

This bring the following advantages: only fetch the remote server if the
key isn't present; report the task as changed only when an actual change
took place.