Create and enable kong service
This is heavily based on the existing `kong` role.
I've tested a curl on the 8001 port (admin) and I receive an apparently
valid answer.

I've run `kong health` in the container and the output is that kong is
healthy. The only that is missing is the config file, which the logs
says will skip loading since it can't find any. The previous
installation did not create a configuration file and it seems that the
default is all we need.

Another reference:
https://hub.docker.com/_/kong