Remove kong-docs service after running it
This is supposed to be a one-shot service. We don't need it on any host
after its execution.

We could use the `docker_compose` ansible module to do that, but we
chose to use the `command` module because the former would require us to
install docker-compose via pip on the gateway machine.