Fix redis url definition on resource adaptor deploy
According to the Sidekiq documentation, the url needs to have the
`redis://` prefix.

But, changing it on the sidekiq initializer results in errors in the CI,
probably because the REDIS_HOST environment variable is being set
somewhere.

This is a solution that makes the deploy work and do not jeopardize the
adaptor's CI.

Reference:
https://github.com/mperham/sidekiq/wiki/Using-Redis#using-an-initializer