Fix kong upstream url
According to the kong documentation, the upstream url is the base target
that points to our API, that will serve for proxying requests. The
kong-api-gateway is the service that provide the proxying rules, but it
is being served by NGINX in the 8888 port, not on the default port (80).

That way, the requests could never reach the destination services.

I took the opportunity to remove the `strip_uri` parameter since the
default is already `true`.

Notice that this is going to be a cause for headaches in the future, as
the API object we are using to register the proxy is deprecated.

Reference:
https://docs.konghq.com/0.13.x/admin-api/