Make kong-docs retry until it succeeds in creating Kong API objects
The `kong-docs` is supposed to be a one-shot service. But it can only
run after both `kong` and `kong-api-gateway` services are ready.

This means that we have to wait to run this container or to retry until
it succeeds. After previous experiences, waiting for dependencies to be
ready (specially more than one) can be hard. So we went with the retry
method that already proved it works.

The downside is that we cannot automatically remove the `kong-docs`
container via ansible commands since we know for sure it won't run
immediately. Since the container is a very small one, we don't think it
is too bad to leave it there and add an instruction for manual removal.