Replace user with become_user on playbooks
We normally do not have ssh access to a root user. Instead, it's better
to have a user who can become a superuser.

This is the difference between the parameters: `user` runs the commands
with the specified user, trying to connect via ssh with this user too.
`become_user` connects with a possibly different user and tries to
change the user to execute a command.