Do not allow services to start if they cannot connect to Kong
There's a explicit dependency among those services and Kong in both
production and development mode.

But, when the services cannot connect to Kong, they do not break. They
simply log the failure and continue.

That causes massive problems for deploying the project. If those
services are started without being able to connect to Kong, they can
never be found. Therefore, currently, for deploying the project we have
to somehow wait for kong to start before starting those services.

That leads to severe complexity and coupling in the deployment scripts.
Here, we prevent the services from starting if they cannot connect to
Kong. This way, those services will not be successfully started and will
have to be retried after some time.