Adjust delay and depends_on parameters on docker compose
The `depends_on` parameter takes no effect on swam mode.
https://docs.docker.com/compose/compose-file/#depends_on-1

Also, we increase the delay on the services that need other services to
run and the ones for tests (hello-world). This was neccessary because
now that the services do not start while Kong is not running, we can end
up in a deadlock and never try to start kong. The `window` was removed
because there is no need to wait to decide if a service restart has
succeeded.