Support standalone installations
There has been a demand to support installations of the platform in a
single instance.

Here we add some configuration to the host file to express that and
adjust the deployment script and docker compose file to accept it.

Some things worth noting:
- The `lookup` function, when passed with a list of a single element,
returns an error. Therefore, we had to add the `wantList` parameter to
it, as suggested by the error message itself.
- The `labels` on the `docker_node` module do not accept booleans as
values, only strings.
- The constraints in the docker compose file work with do not need to
define the values between double or single quotes ('true' or "true").