Create base docker image
This solves the problem of having a `/gems/` directory available for
each service own image. It is important to notice that a Dockerfile
cannot reference a parent directory. Thus `ADD ../gems/ /gems/` is
invalid.

Also we take this opportunity to unify the base image, making easier to
change it and update it, and the installed packages.