Set Ruby and bundler versions on docker images
The applications Gemfile now specify both and after the upgrade they are
different than the ones found at `debian:unstable`. This causes the
service images build to fail:
https://gitlab.com/interscity/interscity-platform/interscity-platform/-/jobs/280504020.

The `ruby` docker images are also debian based. So I do not expect this
base image change to be a breaking one.