Extract Kong wrapper to service-base gem
This class is repeated among all the services. By extracting it to the
service-base gem we'll making easier maintaing it.

For example instead of fixing rubocop issues on 5 files, we'll fix only
on this one.

I'm using the `KongLib` module name in order to avoid the name clash
with the kong gem module that is already named Kong.