Remove unused code from data-collector
It does not contain a module, class or method definition. Also isn't
required anywhere in the code. The commit that adds it has the message
"Fixing bugs".

Thus I consider it as accidentally added and am removing it so we get
rid of unused code that was raising issues on rubocop.