Exclude gemspec files from line length check by Rubocop
The description is supposed to be a long line and there is not much we
can do.

```
gems/service-base/service-base.gemspec:14:121: C: Metrics/LineLength: Line is too long. [209/120]
  spec.description   = 'It is expected to hold gems and configuration files that are shared by more than one service. Its main objective is to eliminate repetitions between services making maintenance easier.'
                                                                                                                        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
```

I choose to make a global exclusion as we intende to have more gems in
this project in the future and they will face the same problem.