Fix missing puma dependency
It seems that the services use Thin as the server when running `rails
s`. This led us to not notice that puma was missing until trying to
deploy to production. Under production, puma is directly invoked to
bring up the server.

This was a mistake introduced by commit
be6a5743f0f8930e2bc76a522b25e3f932c297ee.