Add simple validation step to deploy
After the usual deploy steps, we poll the deployed services' healthcheck
endpoints until they respond with success, so we know the service is up
and able to connect to the databases. Now, as a user deploying the
services, I know the services are up and running as the deploy command
exits, with no extra waiting.