Improve resource data creation documentation
One user understood that the capabilities and their value were part of
the inner most hash in the example. While this inner most hash in the
example is actually a complex document generated by a single capability.

Another problem I found in the previous description was talking about
resources as it was possible to send data for more than one. Which is
not possible as we specify a single resource in the request path URL
attribute.

Rewriting the description I expect to address both.

This also changes the integration test that replicates the mentioned
user's use case. The change fixes it.